Move CommentTracker from transforms/utils to comments directory
Original location: src/plugin/src/transforms/utils/comment-tracker.ts
New location: src/plugin/src/comments/comment-tracker.ts

Rationale:
The CommentTracker class is a comment-handling utility that tracks comment
positions, manages checkpoint/rollback for safe rewrites, and provides
methods to query and manipulate comment ranges. This is fundamentally
comment infrastructure, not a transform-specific utility.

The plugin already has a dedicated src/plugin/src/comments directory for
comment-related functionality (comment-printer, doc-like normalization, etc.).
Moving CommentTracker here aligns with the architectural principle that
comment-handling logic belongs in the comments module, and transforms that
use it are consumers of comment infrastructure, not owners of it.

Changes made:
- Moved comment-tracker.ts from transforms/utils/ to comments/
- Removed now-empty transforms/utils/ directory
- Updated imports in consolidate-struct-assignments.ts
- Updated imports in consolidate-struct-assignment-comment-handler.ts
- Removed CommentTracker export from transforms/index.ts
- Added CommentTracker export to comments/index.ts

All builds and tests pass with no changes in functionality.
