+++ src/plugin/src/comments/comment-tracker.ts
