+++ src/plugin/src/transforms/consolidate-struct-assignments.ts
@@ -4,10 +4,10 @@
  */
 import { Core, type MutableGameMakerAstNode } from "@gml-modules/core";
 
+import { CommentTracker } from "../comments/comment-tracker.js";
 import { AssignmentCommentHandler } from "./consolidate-struct-assignment-comment-handler.js";
 import { type AssignmentDetails, StructAssignmentMatcher } from "./consolidate-struct-assignment-matcher.js";
 import { createParserTransform } from "./functional-transform.js";
-import { CommentTracker } from "./utils/comment-tracker.js";
 
 type CommentTools = {
     addTrailingComment: (...args: Array<unknown>) => unknown;
