+++ src/plugin/src/transforms/index.ts
@@ -87,4 +87,3 @@ export { enforceVariableBlockSpacingTransform } from "./enforce-variable-block-s
 export { applyIndexAdjustmentsIfPresent } from "./index-adjustments.js";
 export { markCallsMissingArgumentSeparatorsTransform } from "./mark-missing-separators.js";
 export { stripCommentsTransform } from "./strip-comments.js";
-export { CommentTracker } from "./utils/comment-tracker.js";
